cirec2/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.0.1
Flask-Mail==0.9.1
Flask-CORS==4.0.0
Werkzeug==2.3.7
WTForms==3.0.1
psycopg2-binary==2.9.7
pgvector==0.2.3
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-dotenv==1.0.0
bcrypt==4.0.1
Pillow==10.0.1
celery==5.3.4
redis==5.0.1
transformers==4.35.2
torch==2.1.0
numpy==1.24.3
pandas==2.0.3
email-validator==2.1.0